Revert "Sub programmer 2"
